+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [v0.0.2] - 2025-10-20
 ### Added
 - Introduced `RegisterFunc` functionality.
+- Add `GetStatus()` for getting current status.
 ### Changed
 - Registration order is now significant.
 - Default WithTimeout is off;